+++ test/components/calendar/calendar.e2e-spec.js
@@ -10,7 +10,7 @@ describe('Calendar index tests', () => {
     await utils.setPage('/components/calendar/example-index?layout=nofrills');
     const dateField = await element(by.css('.calendar-monthview #monthview-datepicker-field'));
     await browser.driver
-      .wait(protractor.ExpectedConditions.presenceOf(dateField), config.waitsFor);
+      .wait(protractor.ExpectedConditions.visibilityOf(dateField), config.waitsFor);
   });
 
   it('Should render without error', async () => {
@@ -19,28 +19,28 @@ describe('Calendar index tests', () => {
   });
 
   it('Should be able to cancel month selector', async () => {
-    expect(await element(by.css('.calendar-monthview #monthview-datepicker-field')).getAttribute('value')).toEqual('November 2019');
+    expect(await element(by.css('.calendar-monthview #monthview-datepicker-field')).getText()).toEqual('November 2019');
 
     await element(by.css('.calendar-monthview #monthview-datepicker-field + .icon'));
     await element(by.css('button.is-cancel'));
 
-    expect(await element(by.css('.calendar-monthview #monthview-datepicker-field')).getAttribute('value')).toEqual('November 2019');
+    expect(await element(by.css('.calendar-monthview #monthview-datepicker-field')).getText()).toEqual('November 2019');
   });
 
   it('Should be able to change month to next', async () => {
     const nextButton = await element(by.css('.calendar-monthview button.next'));
     const testDate = new Date();
 
-    expect(await element(by.css('.calendar-monthview #monthview-datepicker-field')).getAttribute('value')).toEqual(testDate.toLocaleDateString('en-US', { month: 'long', year: 'numeric' }));
+    expect(await element(by.css('.calendar-monthview #monthview-datepicker-field')).getText()).toEqual(testDate.toLocaleDateString('en-US', { month: 'long', year: 'numeric' }));
 
     await nextButton.click();
     await utils.checkForErrors();
     await testDate.setDate(1);
     await testDate.setMonth(testDate.getMonth() + 1);
 
-    expect(await element(by.css('.calendar-monthview #monthview-datepicker-field')).getAttribute('value')).toEqual(testDate.toLocaleDateString('en-US', { month: 'long', year: 'numeric' }));
+    expect(await element(by.css('.calendar-monthview #monthview-datepicker-field')).getText()).toEqual(testDate.toLocaleDateString('en-US', { month: 'long', year: 'numeric' }));
 
-    expect(await element(by.css('.calendar-monthview #monthview-datepicker-field')).getAttribute('value')).toEqual(testDate.toLocaleDateString('en-US', { month: 'long', year: 'numeric' }));
+    expect(await element(by.css('.calendar-monthview #monthview-datepicker-field')).getText()).toEqual(testDate.toLocaleDateString('en-US', { month: 'long', year: 'numeric' }));
     expect(await nextButton.getText()).toEqual('Next Month');
   });
 
@@ -48,7 +48,7 @@ describe('Calendar index tests', () => {
     const prevButton = await element(by.css('.calendar-monthview .btn-icon.prev'));
     const testDate = new Date();
 
-    expect(await element(by.css('.calendar-monthview #monthview-datepicker-field')).getAttribute('value')).toEqual(testDate.toLocaleDateString('en-US', { month: 'long', year: 'numeric' }));
+    expect(await element(by.css('.calendar-monthview #monthview-datepicker-field')).getText()).toEqual(testDate.toLocaleDateString('en-US', { month: 'long', year: 'numeric' }));
     await prevButton.click();
     await utils.checkForErrors();
 
@@ -58,9 +58,38 @@ describe('Calendar index tests', () => {
     await testDate.setMinutes(0);
     await testDate.setSeconds(0);
 
-    expect(await element(by.css('.calendar-monthview #monthview-datepicker-field')).getAttribute('value')).toEqual(testDate.toLocaleDateString('en-US', { month: 'long', year: 'numeric' }));
+    expect(await element(by.css('.calendar-monthview #monthview-datepicker-field')).getText()).toEqual(testDate.toLocaleDateString('en-US', { month: 'long', year: 'numeric' }));
 
     expect(await prevButton.getText()).toEqual('Previous Month');
+  });
+
+  it('Should be able to add with the modal', async () => {
+    const beforeCount = await element.all(by.css('.calendar-monthview .calendar-event')).count();
+    await element.all(by.cssContainingText('.monthview-table td', '13')).first().click();
+    await browser.actions()
+      .doubleClick(await element.all(by.cssContainingText('.monthview-table td', '13')).first())
+      .perform();
+
+    await browser.driver
+      .wait(protractor.ExpectedConditions.visibilityOf(await element(by.id('subject'))), config.waitsFor);
+    await element(by.id('subject')).sendKeys('Test Event');
+    await element(by.id('submit')).click();
+
+    let afterCount = await element.all(by.css('.calendar-event')).count();
+
+    expect(afterCount).toEqual(beforeCount + 2);
+
+    await browser.actions().mouseMove(await element(by.cssContainingText('.calendar-monthview .calendar-event', 'Test Event'))).perform();
+    await browser.actions().click(protractor.Button.RIGHT).perform();
+
+    await browser.driver
+      .wait(protractor.ExpectedConditions.visibilityOf(await element(by.id('calendar-actions-menu'))), config.waitsFor);
+
+    await element.all(by.css('.popupmenu-wrapper li a')).first().click();
+
+    afterCount = await element.all(by.css('.calendar-monthview .calendar-event')).count();
+
+    expect(afterCount).toEqual(beforeCount);
   });
 });
 
@@ -81,7 +110,7 @@ describe('Calendar ajax loading tests', () => {
     await testDate.setMonth(7);
     await testDate.setFullYear(2018);
 
-    expect(await element(by.css('.calendar-monthview #monthview-datepicker-field')).getAttribute('value')).toEqual(testDate.toLocaleDateString('en-US', { month: 'long', year: 'numeric' }));
+    expect(await element(by.css('.calendar-monthview #monthview-datepicker-field')).getText()).toEqual(testDate.toLocaleDateString('en-US', { month: 'long', year: 'numeric' }));
   });
 
   it('Should render ajax loaded dates for august 2018', async () => {
@@ -102,7 +131,7 @@ describe('Calendar ajax loading tests', () => {
     await testDate.setMonth(8);
     await testDate.setFullYear(2018);
 
-    expect(await element(by.css('.calendar-monthview #monthview-datepicker-field')).getAttribute('value')).toEqual(testDate.toLocaleDateString('en-US', { month: 'long', year: 'numeric' }));
+    expect(await element(by.css('.calendar-monthview #monthview-datepicker-field')).getText()).toEqual(testDate.toLocaleDateString('en-US', { month: 'long', year: 'numeric' }));
     expect(await element.all(by.css('.calendar-event-more')).count()).toEqual(0);
     expect(await element.all(by.css('.calendar-event')).count()).toEqual(2);
   });
@@ -125,7 +154,7 @@ describe('Calendar specific month tests', () => {
     await testDate.setMonth(9);
     await testDate.setFullYear(2018);
 
-    expect(await element(by.css('.calendar-monthview #monthview-datepicker-field')).getAttribute('value')).toEqual(testDate.toLocaleDateString('en-US', { month: 'long', year: 'numeric' }));
+    expect(await element(by.css('.calendar-monthview #monthview-datepicker-field')).getText()).toEqual(testDate.toLocaleDateString('en-US', { month: 'long', year: 'numeric' }));
   });
 
   it('Should be able to click on events', async () => {
@@ -183,7 +212,7 @@ describe('Calendar specific month tests', () => {
     expect(await element.all(by.css('.calendar-event')).count()).toEqual(15);
   });
 
-  it('should add new events on click and cancel', async () => {
+  it('Should add new events on click and cancel', async () => {
     expect(await element.all(by.css('.calendar-event')).count()).toEqual(16);
 
     const event = await element.all(by.cssContainingText('.monthview-table td', '1')).first();
@@ -198,7 +227,7 @@ describe('Calendar specific month tests', () => {
     expect(await element.all(by.css('.calendar-event')).count()).toEqual(16);
   });
 
-  it('should add new events on click and submit', async () => {
+  it('Should add new events on click and submit', async () => {
     expect(await element.all(by.css('.calendar-event')).count()).toEqual(16);
 
     const event = await element.all(by.cssContainingText('.monthview-table td', '1')).first();
@@ -214,7 +243,7 @@ describe('Calendar specific month tests', () => {
   });
 });
 
-describe('Calendar only calendar', () => {  //eslint-disable-line
+describe('Calendar only calendar', () => {
   beforeEach(async () => {
     await utils.setPage('/components/calendar/example-only-calendar');
     const dateField = await element(by.css('.calendar-monthview #monthview-datepicker-field'));
@@ -318,7 +347,7 @@ describe('Calendar only monthview and legend', () => {
   }
 });
 
-describe('Calendar WeekView settings tests', () => {  //eslint-disable-line
+describe('Calendar WeekView settings tests', () => {
   beforeEach(async () => {
     await utils.setPage('/components/calendar/test-weekview-settings?layout=nofrills');
 
@@ -327,12 +356,12 @@ describe('Calendar WeekView settings tests', () => {  //eslint-disable-line
       .wait(protractor.ExpectedConditions.visibilityOf(dateField), config.waitsFor);
   });
 
-  it('Should render without error', async () => {  //eslint-disable-line
+  it('Should render without error', async () => {
     expect(await element.all(by.css('.calendar-event')).count()).toEqual(14);
     await utils.checkForErrors();
   });
 
-  it('Should switch to week', async () => {  //eslint-disable-line
+  it('Should switch to week', async () => {
     expect(await element(by.css('.week-view ')).isDisplayed()).toBe(false);
     const dropdownEl = await element.all(by.css('#calendar-view-changer + .dropdown-wrapper div.dropdown')).first();
     await dropdownEl.sendKeys(protractor.Key.ARROW_DOWN);
@@ -347,46 +376,4 @@ describe('Calendar WeekView settings tests', () => {  //eslint-disable-line
 
     expect(await element(by.css('.week-view ')).isDisplayed()).toBe(true);
   });
-});
-
-describe('Calendar Switch to Day view', () => {  //eslint-disable-line
-  beforeEach(async () => {
-    await utils.setPage('/components/calendar/test-specific-month?layout=nofrills');
-    const dateField = await element(by.css('.calendar-monthview #monthview-datepicker-field'));
-    await browser.driver
-      .wait(protractor.ExpectedConditions.visibilityOf(dateField), config.waitsFor);
-  });
-
-  it('Should render without error', async () => {
-    expect(await element.all(by.css('.calendar-event')).count()).toEqual(16);
-    await utils.checkForErrors();
-  });
-
-  if (!utils.isCI()) {
-    it('Should switch to day', async () => {
-      await browser.driver.sleep(config.sleep);
-
-      expect(await element(by.css('.week-view')).isDisplayed()).toBe(false);
-      await element.all(by.cssContainingText('.monthview-table td', '12')).first().click();
-
-      const dropdownEl = await element.all(by.css('#calendar-view-changer + .dropdown-wrapper div.dropdown')).first();
-      await dropdownEl.sendKeys(protractor.Key.ARROW_DOWN);
-
-      const searchEl = await element(by.css('.dropdown-search'));
-      await browser.driver
-        .wait(protractor.ExpectedConditions.presenceOf(searchEl), config.waitsFor);
-
-      await browser.switchTo().activeElement().sendKeys(protractor.Key.ARROW_DOWN);
-      await browser.switchTo().activeElement().sendKeys(protractor.Key.ARROW_DOWN);
-      await browser.switchTo().activeElement().sendKeys(protractor.Key.ENTER);
-
-      await browser.driver
-        .wait(protractor.ExpectedConditions.visibilityOf(await element.all(by.css('.week-view-table .calendar-event')).last()), config.waitsFor);
-      await browser.driver.sleep(config.sleep);
-
-      expect(await element(by.css('.week-view-table .calendar-event')).isDisplayed()).toBe(true);
-      expect(await element.all(by.css('.week-view-table .calendar-event')).count()).toEqual(1);
-      expect(await element(by.css('.week-view-table .calendar-event')).getText()).toEqual('Birthday Off');
-    });
-  }
 });